+++ core/CHANGELOG.md
@@ -1,6 +1,7 @@
 # TBD
 * Add a tutorial explaining what Kurtosis does at the Docker level
 * Kill TODOs in "Debugging Failed Tests" tutorial
+* Build a v0 of Docker container containing the Kurtosis API 
 
 # 0.9.0
 * Change ConfigurationID to be a string